+++ src/schema/index.ts
@@ -9,7 +9,7 @@ import taskType from "./task.graphql";
 import quizTaskSubmission from "./taskSubmission.graphql";
 import questionType from "./question.graphql";
 import taskblockType from "./taskblock.graphql";
-import objectiveType from "./objective.graphql";
+import objectiveType from "../objective/objective.graphql";
 import targetType from "./target.graphql";
 import progressType from "../progress/progress.graphql";
 import goalsType from "./goal.graphql";
@@ -21,7 +21,7 @@ import questionResolvers from "./question.resolver";
 import quizTaskSubmissionResolvers from "./taskSubmission.resolver";
 import missionResolvers from "./mission.resolver";
 import taskblockResolvers from "./taskblock.resolver";
-import objectiveResolvers from "./objective.resolver";
+import objectiveResolvers from "../objective/objective.resolver";
 import targetResolvers from "./target.resolver";
 import progressResolvers from "../progress/progress.resolver";
 import goalsResolvers from "./goal.resolver";
